+++ pydantic/types.py
@@ -1,9 +1,18 @@
 import json
 import re
 from decimal import Decimal
-from ipaddress import IPv4Address, IPv6Address, _BaseAddress
+from ipaddress import (
+    IPv4Address,
+    IPv4Interface,
+    IPv4Network,
+    IPv6Address,
+    IPv6Interface,
+    IPv6Network,
+    _BaseAddress,
+    _BaseNetwork,
+)
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Dict, Generator, Optional, Pattern, Set, Type, Union, cast
+from typing import TYPE_CHECKING, Any, Dict, Generator, Optional, Pattern, Set, Tuple, Type, Union, cast
 from uuid import UUID
 
 from . import errors
@@ -63,6 +72,8 @@ __all__ = [
     'Json',
     'JsonWrapper',
     'IPvAnyAddress',
+    'IPvAnyInterface',
+    'IPvAnyNetwork',
 ]
 
 NoneStr = Optional[str]
@@ -72,6 +83,7 @@ NoneStrBytes = Optional[StrBytes]
 OptionalInt = Optional[int]
 OptionalIntFloat = Union[OptionalInt, float]
 OptionalIntFloatDecimal = Union[OptionalIntFloat, Decimal]
+NetworkType = Union[str, bytes, int, Tuple[Union[str, bytes, int], Union[str, int]]]
 
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -516,3 +528,37 @@ class IPvAnyAddress(_BaseAddress):
 
         with change_exception(errors.IPvAnyAddressError, ValueError):
             return IPv6Address(value)
+
+
+class IPvAnyInterface(_BaseAddress):
+    @classmethod
+    def __get_validators__(cls) -> 'CallableGenerator':
+        yield cls.validate
+
+    @classmethod
+    def validate(cls, value: NetworkType) -> Union[IPv4Interface, IPv6Interface]:
+        try:
+            return IPv4Interface(value)
+        except ValueError:
+            pass
+
+        with change_exception(errors.IPvAnyInterfaceError, ValueError):
+            return IPv6Interface(value)
+
+
+class IPvAnyNetwork(_BaseNetwork):  # type: ignore
+    @classmethod
+    def __get_validators__(cls) -> 'CallableGenerator':
+        yield cls.validate
+
+    @classmethod
+    def validate(cls, value: NetworkType) -> Union[IPv4Network, IPv6Network]:
+        # Assume IP Network is defined with a default value for ``strict`` argument.
+        # Define your own class if you want to specify network address check strictness.
+        try:
+            return IPv4Network(value)
+        except ValueError:
+            pass
+
+        with change_exception(errors.IPvAnyNetworkError, ValueError):
+            return IPv6Network(value)