+++ HISTORY.rst
@@ -3,10 +3,15 @@
 History
 -------
 
+v0.21.1 (unreleased)
+....................
+* add ``IPv{4,6,Any}Network`` and ``IPv{4,6,Any}Interface`` types from ``ipaddress`` stdlib, #333 by @pilosus
+
+
 v0.21.0 (2019-03-15)
 ....................
 * fix typo in ``NoneIsNotAllowedError`` message, #414 by @YaraslauZhylko
-* add ``IPAddress``, ``IPv4Address`` and ``IPv6Address`` types, #333 by @pilosus
+* add ``IPvAnyAddress``, ``IPv4Address`` and ``IPv6Address`` types, #333 by @pilosus
 
 v0.20.1 (2019-02-26)
 ....................